requirements.txt
numpy
scipy
pandas
matplotlib
open3d
pybullet==2.7.9
torch
pytorch-ignite
tensorboard
tqdm
trimesh
torch-scatter
urdfpy
cython
scikit-image